requirements.txt
click==8.1.6
Flask==2.3.2
Flask_Assets==2.0
python-dotenv==1.0.0
